allow float input while performing floor division

+++ print_two_numbers_quotient_without_decimal_point.py
@@ -1,3 +1,3 @@
-first_number = int(input("Enter the first number: "))
-second_number = int(input("Enter the second number: "))
-print(f"The quotient of two numbers is ", first_number // second_number)
+first_number = float(input("Enter the first number: "))
+second_number = float(input("Enter the second number: "))
+print(f"The quotient of two numbers is ", int(first_number // second_number))
